Move migration and backup code to separte files
