+++ lib/core/storage.js
@@ -1,11 +1,8 @@
 const { storage } = require("sdk/simple-storage");
-const base64 = require("sdk/base64");
-const { indexedDB: idb } = require('sdk/indexed-db');
 
+const { indexedDB: idb } = require('sdk/indexed-db');
 const config = require("../config");
-const api_util = require("../api/util");
 const util = require("../util");
-const filters = require("../core/filters");
 
 const max_history_size = 50;
 const required_properties  = ["history", "subscriptions", "videos"];
@@ -58,7 +55,6 @@ const video = {
             }], cb);
         };
     },
-
     count(trans, cb) {
         let req = video_store(trans).count();
         forward_idb_request(req, cb);
@@ -214,114 +210,6 @@ function ensure_valid () {
     });
 }
 
-function export_all () {
-    let channels = [];
-    for (let channel of storage.subscriptions) {
-        channels.push({
-            id: channel.id,
-            title: channel.title,
-            filters: channel.filters || []
-        });
-    }
-    return base64.encode(JSON.stringify({
-        channels: channels,
-        videos: video.get_all()[0],
-        config: config.get_all()
-    }), "utf-8");
-}
-
-function import_all (encoded) {
-    let input;
-    try {
-        input = JSON.parse(base64.decode(encoded, "utf-8"));
-    } catch (e) {
-        return false;
-    }
-    if (!input.hasOwnProperty("channels") || !input.hasOwnProperty("videos") ||
-        !input.hasOwnProperty("config")) {
-        return false;
-    }
-    for (let channel_ of input.channels) {  // validate every filter
-        if (!channel_.filters.every(filters.is_full_filter)) {
-            return false;
-        }
-    }
-    for (let channel_ of input.channels) {
-        let existing_channel = channel.get_by_id(channel_.id);
-        if (existing_channel === undefined) {
-            channel_.latest_date = (new Date()).getTime();
-            storage.subscriptions.push(channel_);
-        } else {  // channel already exists
-            // merge list of filters
-            for (let filter_to_add of channel_.filters) {
-                if (!existing_channel.filters.some(
-                        filters.filters_equal.bind(null, filter_to_add))) {
-                    existing_channel.filters.push(filter_to_add);
-                }
-            }
-        }
-    }
-    for (let video_ of input.videos) {
-        if (video.get_by_id(video_.video_id) === undefined) {
-            storage.videos.push(video_);
-        }
-    }
-    config.update(input.config);
-    return true;
-}
-
-/*
-  This object is responsible for transitioning the old model to the new one
-*/
-const transition = {
-    // change the old video storage model used in 1.0 to the new one
-    // this function assumes the old model is in use.
-    update_storage_model: () => {
-        if (storage.backup) {  // we are downgrading from version 2.0 or above
-            storage.videos = storage.backup;
-            storage.backup = undefined;
-            return;
-        }
-        if (Array.isArray(storage.videos)) {
-            return;
-        }
-        let videos = [];
-        for (let key in storage.videos) {
-            videos = videos.concat(storage.videos[key]);
-        }
-        let durations = [];
-        // copy video duration since api_util.activity.normalize wipes the
-        // property to empty string
-        for (let video of videos) {
-            let video_duration = video.duration;
-            if (video_duration === undefined) {
-                video_duration = "";
-            }
-            durations.push(video_duration);
-        }
-        videos.map(api_util.activity.normalize);
-        for (let i = 0; i < durations.length; i++) {
-            videos[i].duration = durations[i];
-        }
-        storage.videos = videos;
-    },
-    // go back to the old model used by 1.0. This wipes currently unwatched
-    // videos and clear the "video_count" property of every channel object.
-    // The unwatched object would be stored in a property called "backup", so
-    // versions > 1.0 can use that to downgrade more easily.
-    revert_storage_model: () => {
-        storage.backup = storage.videos;
-        storage.videos = {};
-        for (let channel of storage.subscriptions) {
-            channel.video_count = 0;
-        }
-    },
-    // returns whether the legacy (1.0) model is in use
-    legacy_in_use: () => {
-        return !Array.isArray(storage.videos);
-    }
-};
-
 const duration_fetch = {
     add_list(trans, video_ids, cb) {
         let store = duration_fetch_store(trans);
@@ -444,10 +332,7 @@ function initialize_db(cb) {
     };
 }
 
-exports.transition = transition;
 exports.video = video;
-exports.export_all = export_all;
-exports.import_all = import_all;
 exports.channel = channel;
 exports.ensure_valid = ensure_valid;
 exports.open = open;
