+++ lib/main.js
@@ -22,6 +22,7 @@ function get_db() {
 
 const request = require("./api/request");
 const storage = require("./core/storage");
+const migration = require("./core/migration");
 const config = require("./config");
 const filters = require("./core/filters");
 const events = require("./events");
@@ -133,7 +134,6 @@ function process_channel_activities (response_json, cb) {
         // video. One second more since two videos uploaded during the
         // same second is treated as both "later than each other".
         storage.check_stamp.update(trans, channel_id, most_recent + 1000);
-        console.log(new_uploads[0])
         new_uploads.forEach(api_util.activity.normalize);
         [include, exclude] = filters.filter_videos(new_uploads, video_filters);
         new_uploads.forEach(e => delete e.tags);
@@ -261,10 +261,11 @@ function maybe_notify_new_upload(process_result) {
     storage.history.get_all(trans, (_, his) => {
         history = his;
     });
+    const collect_title = (_, channel) => {
+        channel_names.push(channel.title);
+    };
     for (let channel_id of process_result) {
-        storage.channel.get_by_id(trans, channel_id, (_, channel) => {
-            channel_names.push(channel.title);
-        });
+        storage.channel.get_by_id(trans, channel_id, collect_title);
     }
     trans.oncomplete = () => {
         button.update(videos.length);
@@ -410,10 +411,10 @@ if (self.loadReason === "upgrade") {
         events.once_new_target(() => events.notify.open_changelog());
     };
 
-    if (storage.transition.v1InUse()) {
-        storage.transition.v1Tov3(init_and_show_changelog);
-    } else if (storage.transition.v2InUse()) {
-        storage.transition.v2Tov3(init_and_show_changelog);
+    if (migration.v1InUse()) {
+        migration.v1Tov3(init_and_show_changelog);
+    } else if (migration.v2InUse()) {
+        migration.v2Tov3(init_and_show_changelog);
     } else {
         init_and_show_changelog();
     }
